implemented get day by Id

+++ src/main/java/com/akiacevedo/life_line_back_end/web/DayController.java
@@ -28,7 +28,7 @@ public class DayController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Day> getDaysById(@PathVariable int id) {
+    public ResponseEntity<Day> getDaysById(@PathVariable Long id) {
         return ResponseEntity.ok(service.getDayById(id));
     }
